evaluation/pyproject.toml
[project]
name = "codecrdt-evaluation"
version = "1.0.0"
description = "CodeCRDT: Scientific Evaluation Framework for CRDT-based Collaborative Code Generation"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Sergey Pugachev", email = "sergey@spugachev.com" }]
license = { text = "MIT" }
keywords = [
    "CRDT",
    "evaluation",
    "benchmarking",
    "code-generation",
    "collaborative-agents",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "aiohttp>=3.12.15",
    "asyncio-throttle>=1.0.2",
    "boto3>=1.35.0",
    "click>=8.3.0",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "pydantic>=2.11.9",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "scipy>=1.16.2",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "typing-extensions>=4.15.0",
]

[project.scripts]
codecrdt-eval = "evaluation.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/evaluation"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.2",
    "black>=24.12.0",
    "ruff>=0.9.2",
    "mypy>=1.14.1",
    "types-pyyaml>=6.0.12.20250915",
    "pandas-stubs>=2.3.2.250827",
    "scipy-stubs>=1.16.2.0",
    "types-seaborn>=0.13.2.20250914",
]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
